requirements.txt
flask
flask-cors
torch
torchvision
pillow
groq
dotenv
opencv-python
matplotlib